pop up confirmation unfollow

+++ components/ModalFollow.js
@@ -1,4 +1,4 @@
-import { View, Text, StyleSheet, TouchableOpacity, Image } from "react-native";
+import { View, Text, StyleSheet, TouchableOpacity, Alert } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import { FontAwesome6, MaterialIcons } from "@expo/vector-icons";
 import Modal from "react-native-modal";
@@ -11,18 +11,32 @@ export default function ModalFollow({ menuVisible, onClose, username, followedUs
     const navigation = useNavigation();
     const dispatch = useDispatch();
 
-    const handlePress = async () => {
-        const res = await deleteFollowedUser(followedUserId, token);
+    const handleUnfollowPress = async () => {
 
-        if (res.result) {
-            dispatch(unfollowUser({userId: followedUserId}));
-            onClose();
-            navigation.navigate("Abonnements");
-
-        } else {
-            alert("Error while deleting")
-        }
-
+        Alert.alert(
+            "Confirmation", 
+            `Voulez-vous vraiment ne plus suivre ${username} ?`, [
+                {
+                    text: "Annuler",
+                },
+                {
+                    text: "Supprimer",
+                    style: "destructive",
+                    onPress: async () => {
+                        const res = await deleteFollowedUser(followedUserId, token);
+                
+                        if (res.result) {
+                            dispatch(unfollowUser({userId: followedUserId}));
+                            onClose();
+                            navigation.navigate("Abonnements");
+                
+                        } else {
+                            alert("Error while deleting")
+                        }
+                    }
+                }
+            ]
+        );
     };
 
     return (
@@ -44,7 +58,7 @@ export default function ModalFollow({ menuVisible, onClose, username, followedUs
                             ...styles.btnOutline,
                             borderLeftWidth: 0.6,
                         }}
-                        onPress={() => handlePress()}
+                        onPress={() => handleUnfollowPress()}
                     >
                         <Text style={styles.textButton}>Ne plus suivre</Text>
                         <MaterialIcons
